test: move tests and fixtures to top-level "tests" directory
Move test and fixture directories from `__tests__` and `__fixtures__`
to `tests` and `tests/fixtures` respectively. Update import paths,
configuration files, and documentation to reflect the new structure.
