+++ .gitignore
@@ -98,7 +98,7 @@ typings/
 Thumbs.db
 
 # Ignore built ts files
-__tests__/runner/*
+tests/runner/*
 
 # IDE files
 .idea
