+++ tests/fixtures/core.ts
