+++ tsconfig.eslint.json
@@ -7,8 +7,7 @@
   },
   "exclude": ["dist", "node_modules"],
   "include": [
-    "__fixtures__",
-    "__tests__",
+    "tests",
     "src",
     "eslint.config.mjs",
     "jest.config.js",
