+++ tsconfig.json
@@ -6,6 +6,6 @@
     "moduleResolution": "NodeNext",
     "outDir": "./dist"
   },
-  "exclude": ["__fixtures__", "__tests__", "coverage", "dist", "node_modules"],
+  "exclude": ["tests", "coverage", "dist", "node_modules"],
   "include": ["src"]
 }
